Googleログイン認証による高評価動画の取得

+++ app/models/user.rb
@@ -3,6 +3,8 @@ class User < ApplicationRecord
 
   has_many :subscription_channels, dependent: :destroy
   has_many :channels, through: :subscription_channels, source: :channel
+  has_many :popular_videos, dependent: :destroy
+  has_many :videos, through: :popular_videos, source: :video
 
   validates :password, length: { minimum: 3 }, if: -> { new_record? || changes[:crypted_password] }
   validates :password, confirmation: true, if: -> { new_record? || changes[:crypted_password] }
@@ -12,14 +14,17 @@ class User < ApplicationRecord
   validates :reset_password_token, uniqueness: true, allow_nil: true
   validates :remember_me_token, uniqueness: true, allow_nil: true
 
-  def add_subscriptions(subscriptions_channels)
-    channels.delete_all
-    subscriptions_channels.each do |channel|
-      channels << channel unless subscription_channel?(channel)
+  def add_subscriptions(subscription_channels)
+    channels.delete_all if channels.present?
+    subscription_channels.each do |channel|
+      channels << channel unless channels.include?(channel)
     end
   end
 
-  def subscription_channel?(channel)
-    channels.include?(channel)
+  def add_popular_videos(popular_videos)
+    videos.delete_all if videos.present?
+    popular_videos.each do |video|
+      videos << video unless videos.include?(video)
+    end
   end
 end
